+++ alphashape/alphashape.py
@@ -9,9 +9,14 @@ from shapely.ops import cascaded_union, polygonize
 from shapely.geometry import MultiPoint, MultiLineString
 from scipy.spatial import Delaunay
 import numpy as np
+try:
+    import geopandas
+    USE_GP = True
+except ImportError:
+    USE_GP = False
 
 
-def alphashape(points, alpha):
+def alphashape(points, alpha=None):
     """
     Compute the alpha shape (concave hull) of a set of points.  If the number
     of points in the input is three or less, the convex hull is returned to the
@@ -20,21 +25,46 @@ def alphashape(points, alpha):
 
     Args:
 
-        points (list): an iterable container of points
+        points (list or ``shapely.geometry.MultiPoint`` or
+            ``geopandas.GeoDataFrame``): an iterable container of points
 
         alpha (float): alpha value
 
     Returns:
 
         ``shapely.geometry.Polygon`` or ``shapely.geometry.LineString`` or
-        ``shapely.geometry.Point``: the resulting geometry
+        ``shapely.geometry.Point`` or ``geopandas.GeoDataFrame``: the resulting
+            geometry
     """
-    points = MultiPoint(list(points))
+    # If given a geodataframe, extract the geometry
+    if USE_GP and isinstance(points, geopandas.GeoDataFrame):
+        crs = points.crs
+        points = points['geometry']
+    else:
+        crs = None
+
+    if not isinstance(points, MultiPoint):
+        points = MultiPoint(list(points))
 
     # If given a triangle for input, or an alpha value of zero or less,
     # return the convex hull.
-    if len(points) < 4 or alpha <= 0:
-        return points.convex_hull
+    if len(points) < 4 or (alpha is not None and alpha <= 0):
+        result = points.convex_hull
+        if crs:
+            gdf = geopandas.GeoDataFrame(geopandas.GeoSeries(result)).rename(
+                columns={0: 'geometry'}).set_geometry('geometry')
+            gdf.crs = crs
+            return gdf
+        else:
+            return result
+
+    # Determine alpha parameter if one is not given
+    if alpha is None:
+        try:
+            from optimizealpha import optimizealpha
+        except ImportError:
+            from .optimizealpha import optimizealpha
+        alpha = optimizealpha(points)
 
     coords = np.array([point.coords[0] for point in points])
     tri = Delaunay(coords)
@@ -65,6 +95,16 @@ def alphashape(points, alpha):
                     edges.add((i, j))
                     edge_points.append(coords[[i, j]])
 
+    # Create the resulting polygon from the edge points
     m = MultiLineString(edge_points)
     triangles = list(polygonize(m))
-    return cascaded_union(triangles)
+    result = cascaded_union(triangles)
+
+    # Convert to pandas geodataframe object if that is what was an input
+    if crs:
+        gdf = geopandas.GeoDataFrame(geopandas.GeoSeries(result)).rename(
+            columns={0: 'geometry'}).set_geometry('geometry')
+        gdf.crs = crs
+        return gdf
+    else:
+        return result