+++ docs/conf.py
@@ -21,8 +21,8 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.join(os.path.abspath('..'), 'alphashape'))
-from _version import __version__
+sys.path.insert(0, os.path.abspath('..'))
+from alphashape import __version__
 
 
 # -- General configuration ---------------------------------------------
@@ -179,4 +179,5 @@ sphinx_gallery_conf = {
     'line_numbers': True,
     'examples_dirs': '../examples',
     'backreferences_dir': 'gen_modules/backreferences',
+    'doc_module': ('alphashape'),
     'gallery_dirs': 'auto_examples'}