+++ tests/test_alphashape.py
@@ -79,12 +79,17 @@ class TestAlphashape(unittest.TestCase):
             (0., 0.), (0., 1.), (1., 1.), (1., 0.), (0., 0.)]).equals(
                 alphashape([(0., 0.), (0., 1.), (1., 1.), (1., 0.)], 1.e-9))
 
+    def test_given_a_four_point_polygon_with_no_alpha_return_input(self):
+        """
+        Given a polygon with four points, return the input as a polygon.
+        """
+        assert shapely.geometry.Polygon([
+            (0., 0.), (0., 1.), (1., 1.), (1., 0.), (0., 0.)]).equals(
+                alphashape([(0., 0.), (0., 1.), (1., 1.), (1., 0.)]))
+
     def test_command_line_interface(self):
         """Test the CLI."""
         runner = CliRunner()
-        result = runner.invoke(cli.main)
-        assert result.exit_code == 0
-        assert 'alphashape.cli.main' in result.output
         help_result = runner.invoke(cli.main, ['--help'])
         assert help_result.exit_code == 0
-        assert '--help  Show this message and exit.' in help_result.output
+        assert 'Show this message and exit.' in help_result.output