+++ HISTORY.md
@@ -1,16 +1,38 @@
 # History
 
+## 1.0.1 (2019-05-06)
+
+* Added gallery plot for optimized alpha function.
+* Documentation cleanup.
+
+## 1.0.0 (2019-05-06)
+
+* [#1 Update features in README.md](https://github.com/bellockk/alphashape/issues/1)
+* [#2 Create Application Utilizing the alphashape Toolbox](https://github.com/bellockk/alphashape/issues/2)
+
+## 0.1.10 (2019-05-05)
+
+* Correcting formatting on PyPi long description.
+
+## 0.1.9 (2019-05-05)
+
+* [#7 Include GeoPandas Integration](https://github.com/bellockk/alphashape/issues/7)
+
+## 0.1.8 (2019-05-05)
+
+* [#8 Include capability to optimize alpha parameter](https://github.com/bellockk/alphashape/issues/8)
+
 ## 0.1.7 (2019-04-26)
 
 * Complete code coverage of existing capabilities.
 
 ## 0.1.6 (2019-04-24)
 
-* Documentation cleanup.
+* [#6 Include Jupyter Notebook in Examples](https://github.com/bellockk/alphashape/issues/6)
 
 ## 0.1.5 (2019-04-24)
 
-* Documentation cleanup.
+* [#5 Create an Example Gallery in the Documentation](https://github.com/bellockk/alphashape/issues/5)
 
 ## 0.1.4 (2019-04-24)